Stabilize Tahoe audio routing around aggregate devices and shutdown
Recent post-v0.4.3 Tahoe reports pointed to two failure modes that made
Background Music feel unsafe to use: aggregate output devices could lose
usable volume controls, and quitting or uninstalling could leave the
system stuck at the wrong volume. This change pulls the minimal proven
fixes onto a single branch so Tahoe users can build from source without
reopening the broader helper/ducking design.

The aggregate-device lane now treats aggregate outputs as special cases
for control detection and skips deprecated volume/mute sync paths that
can corrupt System Settings state. The termination lane deactivates
control-sync and playthrough before restoring the default device, and the
uninstall path removes stale BGM device preference entries that can keep
incorrect volume state alive across restarts.

Constraint: Must stay close to upstream master to preserve CoreAudio behavior and ease future rebases
Constraint: Cannot fully verify privileged install/runtime helper behavior in this environment because sudo is unavailable
Rejected: Large Tahoe-specific refactor of the helper/ducking pipeline | too broad without a reproducible runtime failure locally
Rejected: Leaving the open Tahoe stability fixes unmerged locally | keeps known aggregate/shutdown regressions in the user build
Confidence: medium
Scope-risk: moderate
Reversibility: clean
Directive: Treat this as a Tahoe stability lane, not a complete fix for FaceTime/browser ducking; validate runtime capture separately on the target machine
Tested: xcodebuild build (Background Music Device, Background Music, BGMXPCHelper)
Tested: xcodebuild test -only-testing:BGMDriverTests
Tested: xcodebuild test -only-testing:BGMAppUnitTests
Not-tested: Privileged install of BGMXPCHelper/driver via launchd
Not-tested: Real FaceTime + browser YouTube runtime on macOS Tahoe 26.3.1
Related: upstream PR #849
Related: upstream PR #850

+++ BGMDriver/BGMDriver/BGM_NullDevice.cpp
@@ -471,8 +471,8 @@ void    BGM_NullDevice::GetZeroTimeStamp(Float64& outSampleTime,
         (static_cast<Float64>(mNumberTimeStamps) * theHostTicksPerPeriod);
 
     // Set the return values.
-    outSampleTime = mNumberTimeStamps * kZeroTimeStampPeriod;
-    outHostTime = static_cast<UInt64>(mAnchorHostTime + theHostTicksSinceAnchor);
+    outSampleTime = static_cast<Float64>(mNumberTimeStamps) * kZeroTimeStampPeriod;
+    outHostTime = static_cast<UInt64>(static_cast<Float64>(mAnchorHostTime) + theHostTicksSinceAnchor);
     outSeed = 1;
 }
